types(Coverage): Organize the types source to match the project structure
