+++ components/ForTenantsInterface/components/PlanIdInterface/components/ProductsStep/components/AddProductsContent/AddProductsContent.tsx
@@ -1,39 +1,10 @@
-import { TenantProduct } from "@/context/TenantContext/types";
 import { AddProductsInterface } from "../AddProductsInterface";
-import { Dispatch, SetStateAction } from "react";
+import { AddProductsContentProps } from "../../../../types";
 
-const AddProductsContent = ({
+const AddProductsContent: React.FC<AddProductsContentProps> = ({
   selectedPlan,
   productsImport,
   importProducts,
-}: {
-  selectedPlan: {
-    id: string;
-    name: string;
-    price: string;
-  };
-  productsImport: {
-    fileName: string | null;
-    rawProducts: [][];
-    products: TenantProduct[];
-    errors: string[];
-    isValidated: boolean;
-    handleFileUpload: (file: File) => Promise<void>;
-    validateProducts: () => boolean;
-    updateProducts: Dispatch<SetStateAction<TenantProduct[]>>;
-    setPage: Dispatch<SetStateAction<number>>;
-    paginatedProducts: TenantProduct[];
-    page: number;
-    pagination: {
-      totalItems: number;
-      perPage: number;
-      currentPage: number;
-      totalPages: number;
-      hasNext: boolean;
-      hasPrev: boolean;
-    };
-  };
-  importProducts: (products: TenantProduct[]) => Promise<void>;
 }) => {
   return (
     <>
