+++ components/ForTenantsInterface/components/PlanIdInterface/components/ProductsStep/components/AddProductsInterface/components/SaveProductsButton/SaveProductsButton.tsx
@@ -1,7 +1,4 @@
-type SaveProductsButtonProps = {
-  onClick: () => void;
-};
-
+import { SaveProductsButtonProps } from "@/components/ForTenantsInterface/components/PlanIdInterface/types";
 const SaveProductsButton = ({ onClick }: SaveProductsButtonProps) => {
   return (
     <button onClick={onClick} className="bg-indigo-500 py-3 rounded-lg">
