+++ hooks/types/index.ts
@@ -1,0 +1,31 @@
+import { Tenant, TenantRegisterStep } from "@/context/TenantContext/types";
+
+type RawProductsProps = [];
+
+type StepHeaderProps = {
+  stepNumber: number;
+  stepText: string;
+};
+
+type StepHeaderInterfaceProps = {
+  stepHeader: StepHeaderProps;
+  planId: string;
+  step: TenantRegisterStep;
+};
+
+interface UseTenantsReturn {
+  tenants: Tenant[];
+  filtered: Tenant[];
+  loading: boolean;
+  error: string | null;
+  search: string;
+  setSearch: (value: string) => void;
+  refetch: () => Promise<void>;
+}
+
+export type {
+  RawProductsProps,
+  StepHeaderProps,
+  UseTenantsReturn,
+  StepHeaderInterfaceProps,
+};
